+++ src/Relude/Base.hs
@@ -14,11 +14,7 @@ Reexports from @Data.*@ and @GHC.*@ modules of
 
 module Relude.Base
        ( -- * Base types
-         module Data.Bits
-       , module Data.Char
-       , module Data.Int
-       , module Data.Word
-       , Natural
+         module Data.Char
 
          -- * Base type classes
        , module Data.Eq
@@ -36,12 +32,10 @@ module Relude.Base
 
        , module GHC.Base
        , module GHC.Enum
-       , module GHC.Float
        , module GHC.Generics
-       , module GHC.Num
-       , module GHC.Real
        , module GHC.Show
 
+-- * GHC-specific functionality
 #if MIN_VERSION_base(4,10,0)
        , module GHC.TypeNats
 #else
@@ -54,11 +48,7 @@ module Relude.Base
        ) where
 
 -- Base types
-import Data.Bits (xor)
 import Data.Char (Char, chr)
-import Data.Int (Int, Int16, Int32, Int64, Int8)
-import Data.Word (Word, Word16, Word32, Word64, Word8, byteSwap16, byteSwap32, byteSwap64)
-import Numeric.Natural (Natural)
 
 -- IO
 import System.IO (FilePath, Handle, IO, IOMode (..), stderr, stdin, stdout, withFile)
@@ -74,13 +64,9 @@ import Data.Proxy (Proxy (..))
 import Data.Typeable (Typeable)
 import Data.Void (Void, absurd, vacuous)
 
-import GHC.Base (String, asTypeOf, maxInt, minInt, ord, seq, ($!))
+import GHC.Base (String, asTypeOf, ord, seq, ($!))
 import GHC.Enum (Bounded (..), Enum (..), boundedEnumFrom, boundedEnumFromThen)
-import GHC.Float (Double (..), Float (..), Floating (acos, acosh, asin, asinh, atan, atanh, cos, cosh, exp, logBase, pi, sin, sinh, sqrt, tan, tanh, (**)))
 import GHC.Generics (Generic)
-import GHC.Num (Integer, Num (..), subtract)
-import GHC.Real (Fractional (..), Integral (..), Ratio, Rational, Real (..), RealFrac (..),
-                 denominator, even, fromIntegral, gcd, lcm, numerator, odd, realToFrac, (^), (^^))
 import GHC.Show (Show)
 
 #if MIN_VERSION_base(4,10,0)